ANDROID: Add a boringssl variant of the openssl code
Submitting this back upstream is TODO. Ideally it should be possible to
change the openssl version upstream to be boringssl compatible.

Implemented using the boringssl porting guide:
https://boringssl.googlesource.com/boringssl/+show/471abb1/PORTING.md

There are two main changes:
 - EVP_PKEY_HMAC is replaced with equivalent HMAC_* operations
 - TPm channel encryption using AES-CFB is removed.

BoringSSL silently dropped AES-CFB in the first version and later added
it back to its "decrepit" library, but I have not yet found an
authoritative reason why it was deleted. As AES-CFB is baked into the
TPM standard, this change in boringssl makes the tpm2-tss implementation
strictly worse.

It may be possible to make a case for putting tpm2-tss on an allow-list
for boringssl decrepit, depending on the reason why it was originally
removed.

Bug: 148102516
Test: New code is not built in this commit.
Change-Id: I80c123e2f457dce3c83171f4861d3d0c04121d07